requirements.txt
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
newsapi-python==0.2.7
protobuf==3.20.3
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
